requirements.txt
Flask
pandas
numpy
scikit-learn
matplotlib
seaborn
mysql-connector-python
